svg.editor: implementated go to referenced node with ctrl+click
Implemented the same behaviour as in MPS. When executing a control click
on a reference cell we either select the target node when it exists in the
current editor, or just open its containing root in a new one.

Changed the key mapping for meta key. The used code was specific for firefox.
Now it should work with Safari and Chrome.