Filter out rapid changes on the cloud model
Changes of the cloud model are scheduled for execution on the
event thread. Sometimes the cloud model changes so fast
that there is already a new version before the synchronization
of the previous version even started. These changes are now skiped
and only one sync is executed for multiple changes.