/generatorOutput now runs asynchronous and also caches the result
Some models take more than 30 seconds to generate and then the http call
runs into an timeout.
The model hash calculated by MPS on the model file is used to check if
the cached generator output is still valid. MPS uses the same hash to
show "generation required" in the logical view.